fastapi-server/pyproject.toml
[project]
name = "mycastle-fastapi-mcp"
version = "1.0.0"
description = "FastAPI MCP Server for MyCastle - ESL School Management"
authors = [{name = "MyCastle Team"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "mcp>=1.1.0",
    "supabase>=2.3.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pyjwt>=2.8.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.1",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "email-validator>=2.1.0",
    "orjson>=3.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
